Small refactor of reactrs
